main/web/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
fonttools==4.38.0
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.23.5
packaging==21.3
Pillow==9.3.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
rptree==0.1.1
six==1.16.0
SQLAlchemy==1.4.44
urllib3==1.26.12
Werkzeug==2.2.2
